完成图片上传功能

+++ src/main/java/com/hc/dao/PicTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PicTypeRepository extends CrudRepository<PicType, Integer> {
+public interface PicTypeRepository extends CrudRepository<PicType, Long> {
 
 }
